refactor(spindle-ui): 共通処理をhooksとして切り出し
