+++ packages/spindle-ui/src/NavigationTab/InlineTab/InlineTab.tsx
@@ -1,11 +1,5 @@
-import React, {
-  createRef,
-  type RefObject,
-  useCallback,
-  useEffect,
-  useRef,
-  useState,
-} from 'react';
+import React from 'react';
+import { useTabList } from '../hooks';
 
 type Option = {
   label: string;
@@ -30,70 +24,11 @@ export const InlineTab: React.FC<Props> = ({
   options,
   onClick,
 }) => {
-  const [selectedId, setSelectedId] = useState(defaultSelectedId);
-
-  const buttonsRef = useRef<RefObject<HTMLButtonElement | null>[]>([]);
-
-  const handleClick = useCallback(
-    (
-      e:
-        | React.MouseEvent<HTMLButtonElement>
-        | React.KeyboardEvent<HTMLButtonElement>,
-      id: string,
-    ) => {
-      // 既に選択中の項目に対してクリックした場合は何もしない
-      if (id === selectedId) {
-        return;
-      }
-
-      setSelectedId(id);
-
-      if (typeof onClick === 'function') {
-        onClick(e, id);
-      }
-    },
-    [onClick, selectedId],
-  );
-
-  const handleKeyDown = useCallback(
-    (e: React.KeyboardEvent<HTMLButtonElement>, index: number) => {
-      const totalLength = options.length;
-
-      switch (e.key) {
-        case 'ArrowLeft': {
-          // 基本的には1つ前の要素に移動。ただし、既に先頭の要素にいる場合はリストの最後尾に移動
-          const prevButtonIndex = index - 1 < 0 ? totalLength - 1 : index - 1;
-          const prevButtonRef = buttonsRef.current[prevButtonIndex];
-          prevButtonRef.current?.focus();
-          handleClick(e, options[prevButtonIndex].id);
-          break;
-        }
-        case 'ArrowRight': {
-          // 基本的には1つ後の要素に移動。ただし、既に最後尾の要素にいる場合はリストの先頭に移動
-          const nextButtonIndex = index + 1 >= totalLength ? 0 : index + 1;
-          const nextButtonRef = buttonsRef.current[nextButtonIndex];
-          nextButtonRef.current?.focus();
-          handleClick(e, options[nextButtonIndex].id);
-          break;
-        }
-        case 'Enter': {
-          const targetButton = buttonsRef.current[index].current;
-          // 既に選択中の項目に対してEnterを押下した場合は無効にする
-          if (targetButton?.getAttribute('aria-selected') === 'true') {
-            e.preventDefault();
-          }
-          break;
-        }
-      }
-    },
-    [options, handleClick],
-  );
-
-  useEffect(() => {
-    buttonsRef.current = options.map(
-      (_, index) => buttonsRef.current[index] ?? createRef<HTMLButtonElement>(),
-    );
-  }, [options]);
+  const { selectedId, buttonsRef, handleClick, handleKeyDown } = useTabList({
+    defaultSelectedId,
+    options,
+    onClick,
+  });
 
   if (options.length === 0) {
     return null;
@@ -116,7 +51,7 @@ export const InlineTab: React.FC<Props> = ({
             type="button"
             ref={buttonsRef.current[index]}
             role="tab"
-            onClick={(e) => handleClick(e, id)}
+            onClick={(e) => handleClick(e, id, index)}
             onKeyDown={(e) => handleKeyDown(e, index)}
           >
             <span className={`${BLOCK_NAME}-label`}>{label}</span>
